requirements.txt
pyyaml
imageio
numpy>=1.11.1
matplotlib==3.6.1
scipy
tqdm
wandb
hydra-core
pytorch_lightning==1.7
torchmetrics
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
-f https://data.pyg.org/whl/torch-1.12.0+cu116.html
ipdb
pdbpp
networkx==2.8.7
#-e ggg_data
overrides
pygsp
pyemd